+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 v0.10.0
 ------------------------------
+*February 7, 2018*
+
+### Added
+- Migrated 'conditionalRequired' rule from other project
+
+
+v0.10.0
+------------------------------
 *February 6, 2018*
 
 ### Added